chore: relocate API hook files for better structure
