+++ src/services/home/query/useGetPopularItem.ts
@@ -1,17 +1,15 @@
-import { getCloseDeadlineItem } from '@/api/sellerItem/handleHomeItemList.api';
+import { getPopularItem } from '@/api/home/handleHomeItemList.api';
 import { QUERY_KEYS } from '@/constants/api';
 import { useInfiniteQuery } from '@tanstack/react-query';
-export const useGetCloseDeadlineItem = ({
-  size = 10,
-}: { size?: number } = {}) => {
+export const useGetPopularItem = ({ size = 10 }: { size?: number }) => {
   const query = useInfiniteQuery({
-    queryKey: [QUERY_KEYS.HOME_CLOSE_DEADLINE, size],
+    queryKey: [QUERY_KEYS.HOME_POPULAR, size],
     staleTime: 10 * 60 * 1000,
     gcTime: 15 * 60 * 1000,
     queryFn: async ({ pageParam = 1, queryKey }) => {
       const [, size] = queryKey as [string, number];
 
-      return getCloseDeadlineItem({
+      return getPopularItem({
         page: pageParam,
         size,
       });
